+++ controllers/ProductController.php
@@ -105,6 +105,10 @@ class ProductController extends Controller
      */
     public function actionCreate()
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
         $model = new Product();
 			$model->setAttribute('mfr', '-');
 			$model->setAttribute('model', '-');
@@ -122,6 +126,10 @@ class ProductController extends Controller
 	
 	public function actionCreateCut()
 	{
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
        $model = new Product();
 			$model->setAttribute('mfr', '0');
 			$model->setAttribute('model', '0');
@@ -139,6 +147,10 @@ class ProductController extends Controller
 
 	public function actionCopy()
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
 		$id = Yii::$app->getRequest()->getQueryParam('id');
 		$product = Product::findOne($id);
 		$product_field = ProductField::find()->where(['product_id' => $id])->all();
@@ -197,6 +209,10 @@ class ProductController extends Controller
      */
     public function actionUpdate($id)
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
         $model = $this->findModel($id);
 		
 		if($model->getAttribute('cut_price') == null){
@@ -216,6 +232,10 @@ class ProductController extends Controller
      */
     public function actionDelete($id)
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
 		$product = Product::findOne($id);
         if($product->cut_price > 0){
 			$this->findModel($id)->delete();			
@@ -254,9 +274,42 @@ class ProductController extends Controller
 	
 	public function actionToCart($id)
 	{
-		$model = Product::findOne($id);
+		$model = Product::findOne($id);		
+		$model->product_field = ProductField::find()->where(['product_id' => $model->getAttribute('id')])->all();			
+		$model->field_type[] = FieldType::find()->orderBy('order_by')->all();
+		foreach ($model->product_field as $pf) {
+			$field = Field::find()->where(['id' => $pf->getAttribute('field_id')])->all();
+			$model->field[] = $field;
+		}
+		
 		$cart = \Yii::$app->cart;
 		$cart->add($model);
 		return count($cart->getItems());
+	}
+	
+	public function actionToComparison($id)
+	{
+		$model = Product::findOne($id);		
+		$model->product_field = ProductField::find()->where(['product_id' => $model->getAttribute('id')])->all();			
+		$model->field_type[] = FieldType::find()->orderBy('order_by')->all();
+		foreach ($model->product_field as $pf) {
+			$field = Field::find()->where(['id' => $pf->getAttribute('field_id')])->all();
+			$model->field[] = $field;
+		}
+		
+		$comparison = \Yii::$app->comparison;
+		$comparison->add($model);
+		return count($comparison->getItems());
+	}
+	
+	public function IsAdmin()
+	{
+		$identity = Yii::$app->user->identity;
+		$is_admin = false;
+		if(isset($identity))
+		{
+			$is_admin = $identity->isAdmin;	
+		}
+		return $is_admin;
 	}
 }
